@@ -75,17 +88,16 @@
     "@types/node": "^16.11.12",
     "@types/semver": "^7.3.9",
     "cookie": "^0.5.0",
-    "firebase": "^9.8.1",
-    "firebase-admin": "^10.1.0",
-    "firebase-functions": "^3.20.1",
+    "firebase": "^9.9.0",
+    "firebase-admin": "^11.0.0",
+    "firebase-functions": "^3.22.0",
     "inquirer": "^8.2.0",
     "lru-cache": "^7.8.1",
     "next": "^12.0.7",
     "nuxt": "^2.15.8",
-    "nuxt3": "^3.0.0-27474800.8dd77d7",
-    "replace-in-file": "^6.3.2",
+    "nuxt3": "^3.0.0-rc.6-27635653.88f4525",
     "rimraf": "^3.0.2",
-    "ts-node": "^10.7.0",
-    "typescript": "^4.5.3"
+    "ts-node": "^10.9.1",
+    "typescript": "^4.7.4"
   }
 }
