@@ -1,15 +1,28 @@
-import { NodeJsAsyncHost } from '@angular-devkit/core/node';
-import { workspaces, logging } from '@angular-devkit/core';
-import { WorkspaceNodeModulesArchitectHost } from '@angular-devkit/architect/node';
-import { Target, Architect, targetFromTargetString, targetStringFromTarget } from '@angular-devkit/architect';
+import type { Target } from '@angular-devkit/architect';
 import { mkdir } from 'fs/promises';
 import { join } from 'path';
 import { copy } from 'fs-extra';
 import { parse } from 'jsonc-parser';
+import { pathToFileURL } from 'url';
 
-import { DeployConfig, findDependency, PathFactory, spawn } from '../../utils';
+// Used by the build process, don't shake
+const _pathToFileUrl = pathToFileURL;
+
+import { Commands, DeployConfig, findDependency, PathFactory, spawn } from '../../utils.js';
+
+class MyError extends Error {
+    constructor(reason: string) {
+        console.error(reason);
+        super();
+    }
+}
 
 export const build = async (config: DeployConfig | Required<DeployConfig>, getProjectPath: PathFactory) => {
+
+    const { NodeJsAsyncHost } = await import('@angular-devkit/core/node/index.js');
+    const { workspaces, logging } = await import('@angular-devkit/core/src/index.js');
+    const { WorkspaceNodeModulesArchitectHost } = await import('@angular-devkit/architect/node/index.js');
+    const { Architect, targetFromTargetString, targetStringFromTarget } = await import('@angular-devkit/architect/src/index.js');
 
     // TODO log to firebase-tools
     const logger = new logging.Logger('firebase-tools');
@@ -26,14 +39,24 @@ export const build = async (config: DeployConfig | Required<DeployConfig>, getPr
     let prerenderTarget: Target|undefined;
 
     if (!project) {
-        const angularJson = parse(await host.readFile('angular.json'));
+        const angularJson = parse(await host.readFile(getProjectPath('angular.json')));
         project = angularJson.defaultProject;
-        if (!project) throw `angular.json missing defaultProject`;
     }
-    // TODO if there are multiple projects warn
+
+    if (!project) {
+        const apps: string[] = [];
+        workspace.projects.forEach((value, key) => {
+            if (value.extensions.projectType === 'application') apps.push(key);
+        });
+        if (apps.length === 1) project = apps[0];
+    }
+
+    if (!project) throw new MyError('Unable to detirmine the application to deploy. Use the NG_DELPOY_PROJECT enivornment varaible or `ng deploy` via @angular/fire.');
+
     const workspaceProject = workspace.projects.get(project);
-    if (!workspaceProject) throw `No project ${project} found.`;
+    if (!workspaceProject) throw new MyError(`No project ${project} found.`);
     const deployTargetDefinition = workspaceProject.targets.get('deploy');
+
     if (deployTargetDefinition?.builder === '@angular/fire:deploy') {
         const options = deployTargetDefinition.options;
         if (typeof options?.prerenderTarget === 'string')
@@ -42,33 +65,46 @@ export const build = async (config: DeployConfig | Required<DeployConfig>, getPr
             browserTarget = targetFromTargetString(options.browserTarget);
         if (typeof options?.serverTarget === 'string')
             serverTarget = targetFromTargetString(options.serverTarget);
+        if (!browserTarget) throw new MyError('ng-deploy is missing a browser target. Plase check your angular.json.');
         if (prerenderTarget) {
             const prerenderOptions = await architectHost.getOptionsForTarget(prerenderTarget);
-            if (browserTarget) {
-                if (targetStringFromTarget(browserTarget) !== prerenderOptions?.browserTarget)
-                    throw 'foo';
-            } else {
-                if (typeof prerenderOptions?.browserTarget !== 'string') throw 'foo';
-                browserTarget = targetFromTargetString(prerenderOptions.browserTarget);
-            }
+            if (targetStringFromTarget(browserTarget) !== prerenderOptions?.browserTarget)
+                throw new MyError('ng-deploy\'s browserTarget and prerender\'s browserTarget do not match. Please check your angular.json');
             if (serverTarget && targetStringFromTarget(serverTarget) !== prerenderOptions?.serverTarget)
-                throw 'foo';
+                throw new MyError('ng-deploy\'s serverTarget and prerender\'s serverTarget do not match. Please check your angular.json');
+            if (!serverTarget) console.warn('Treating the application as fully rendered. Add a serverTarget to your deploy target in angular.json to utilize server-side rendering.');
         }
     } else if (workspaceProject.targets.has('prerender')) {
-        // TODO test and warn if production doesn't exist, fallback to default
-        prerenderTarget = { project, target: 'prerender', configuration: 'production' };
+        const target = workspaceProject.targets.get('prerender')!;
+        const configurations = Object.keys(target.configurations!);
+        const configuration = configurations.includes('production') ? 'production' : target.defaultConfiguration;
+        if (!configuration) throw new MyError('No production or default configutation found for prerender.');
+        if (configuration !== 'production') console.warn(`Using ${configuration} configuration for the prerender, we suggest adding a production target.`);
+        prerenderTarget = { project, target: 'prerender', configuration };
         const production = await architectHost.getOptionsForTarget(prerenderTarget);
-        if (typeof production?.browserTarget !== 'string') throw 'foo';
+        if (typeof production?.browserTarget !== 'string')
+            throw new MyError('Prerender browserTarget expected to be string, check your angular.json.');
         browserTarget = targetFromTargetString(production.browserTarget);
-        if (typeof production?.serverTarget !== 'string') throw 'foo';
+        if (typeof production?.serverTarget !== 'string')
+            throw new MyError('Prerender serverTarget expected to be string, check your angular.json.');
         serverTarget = targetFromTargetString(production.serverTarget);
     } else {
-        // TODO test and warn if production doesn't exist, fallback to default
-        const configuration = 'production';
-        if (workspaceProject.targets.has('build'))
+        if (workspaceProject.targets.has('build')) {
+            const target = workspaceProject.targets.get('build')!;
+            const configurations = Object.keys(target.configurations!);
+            const configuration = configurations.includes('production') ? 'production' : target.defaultConfiguration;
+            if (!configuration) throw new MyError('No production or default configutation found for build.');
+            if (configuration !== 'production') console.warn(`Using ${configuration} configuration for the browser deploy, we suggest adding a production target.`);
             browserTarget = { project, target: 'build', configuration };
-        if (workspaceProject.targets.has('server'))
+        }
+        if (workspaceProject.targets.has('server')) {
+            const target = workspaceProject.targets.get('server')!;
+            const configurations = Object.keys(target.configurations!);
+            const configuration = configurations.includes('production') ? 'production' : target.defaultConfiguration;
+            if (!configuration) throw new MyError('No production or default configutation found for server.');
+            if (configuration !== 'production') console.warn(`Using ${configuration} configuration for the server deploy, we suggest adding a production target.`);
             serverTarget = { project, target: 'server', configuration };
+        }
     }
 
     const scheduleTarget = async (target: Target) => {
@@ -77,21 +113,18 @@ export const build = async (config: DeployConfig | Required<DeployConfig>, getPr
         if (!success) throw new Error(error);
     }
 
-    if (!browserTarget) throw 'No build target...';
+    if (!browserTarget) throw new MyError('No build target...');
 
     if (prerenderTarget) {
-        // TODO fix once we can migrate to ESM. Spawn for now.
-        // ERR require() of ES Module .../node_modules/@nguniversal/express-engine/fesm2015/express-engine.mjs not supported.
-        //     Instead change the require of .../node_modules/@nguniversal/express-engine/fesm2015/express-engine.mjs to a dynamic
-        //     import() which is available in all CommonJS modules.
+        // TODO there is a bug here. Spawn for now.
         // await scheduleTarget(prerenderTarget);
         await spawn(
-            'node_modules/.bin/ng',
+            Commands.ng,
             ['run', targetStringFromTarget(prerenderTarget)],
-            { cwd: process.cwd() },
+            { cwd: getProjectPath() },
             // TODO log to firebase-tools
-            out => console.log(out.toString()),
-            err => console.error(err.toString())
+            (out: any) => console.log(out.toString()),
+            (err: any) => console.error(err.toString())
         );
     } else {
         await scheduleTarget(browserTarget);
@@ -102,7 +135,7 @@ export const build = async (config: DeployConfig | Required<DeployConfig>, getPr
     const getHostingPath = (...args: string[]) => deployPath('hosting', ...args);
 
     const browserTargetOptions = await architectHost.getOptionsForTarget(browserTarget);
-    if (typeof browserTargetOptions?.outputPath !== 'string') throw 'foo';
+    if (typeof browserTargetOptions?.outputPath !== 'string') throw new MyError('browserTarget output path is not a string');
     const browserOutputPath = browserTargetOptions.outputPath;
     await mkdir(getHostingPath(), { recursive: true });
     await copy(getProjectPath(browserOutputPath), getHostingPath());
@@ -110,10 +143,10 @@ export const build = async (config: DeployConfig | Required<DeployConfig>, getPr
     const usingCloudFunctions = !!serverTarget;
 
     let bootstrapScript = '';
-    const packageJson = JSON.parse(await host.readFile('package.json'));
+    const packageJson = JSON.parse(await host.readFile(getProjectPath('package.json')));
     if (serverTarget) {
         const serverTargetOptions = await architectHost.getOptionsForTarget(serverTarget);
-        if (typeof serverTargetOptions?.outputPath !== 'string') throw 'foo';
+        if (typeof serverTargetOptions?.outputPath !== 'string') throw new MyError('serverTarget output path is not a string');
         const serverOutputPath = serverTargetOptions.outputPath;
         await mkdir(deployPath('functions', serverOutputPath), { recursive: true });
         await mkdir(deployPath('functions', browserOutputPath), { recursive: true });
