@@ -16,7 +16,8 @@ import { readFile, mkdir } from 'fs/promises';
 import { join } from 'path';
 import { copy } from 'fs-extra';
 
-import { DeployConfig, PathFactory, exec, spawn } from '../../utils';
+import { DeployConfig, PathFactory, exec, spawn, Commands } from '../../utils.js';
+import { pathToFileURL } from 'url';
 
 export const build = async (config: DeployConfig | Required<DeployConfig>, getProjectPath: PathFactory) => {
 
@@ -25,9 +26,9 @@ export const build = async (config: DeployConfig | Required<DeployConfig>, getPr
 
     if (packageJson.scripts?.build) {
         // TODO add to the firebaseTools logs
-        await spawn('npm', ['run', 'build'], { cwd: getProjectPath() }, stdoutChunk => {
+        await spawn(Commands.NPM, ['run', 'build'], { cwd: getProjectPath() }, (stdoutChunk: any) => {
             console.log(stdoutChunk.toString());
-        }, errChunk => {
+        }, (errChunk: any) => {
             console.error(errChunk.toString());
         });
     }
@@ -37,11 +38,11 @@ export const build = async (config: DeployConfig | Required<DeployConfig>, getPr
         const allowRecursion = !entry;
         entry ||= await (async () => {
             try {
-                const requiredProject = require(getProjectPath());
+                const requiredProject = require(pathToFileURL(getProjectPath()).toString());
                 if (requiredProject) method = ['require'];
                 return requiredProject;
             } catch(e) {
-                const importedProject = await import(getProjectPath()).catch(() => undefined);
+                const importedProject = await import(pathToFileURL(getProjectPath()).toString()).catch(() => undefined);
                 if (importedProject) method = ['import'];
                 return importedProject;
             }
