@@ -1,8 +1,10 @@
-import { Request as FunctionsRequest, onRequest } from 'firebase-functions/v2/https';
+import { Request as FunctionsRequest } from 'firebase-functions/v2/https';
 import type { Response } from 'express';
 import { initializeApp as initializeAdminApp } from 'firebase-admin/app';
 import { getAuth as getAdminAuth } from 'firebase-admin/auth';
+// @ts-ignore
 import { initializeApp, deleteApp, FirebaseApp } from 'firebase/app';
+// @ts-ignore
 import { getAuth, signInWithCustomToken, User } from 'firebase/auth';
 import cookie from 'cookie';
 import LRU from 'lru-cache';
@@ -12,11 +14,9 @@ import {
     ID_TOKEN_MAX_AGE,
     LRU_MAX_INSTANCES,
     LRU_TTL
-} from '../constants';
+} from '../constants.js';
 
-const { FRAMEWORK, HTTPS_OPTIONS } = require(`${process.cwd()}/settings`);
-const { handle: frameworkHandle } = require(`../frameworks/${FRAMEWORK}/server`);
-const FIREBASE_PROJECT_CONFIG = process.env.FRAMEWORKS_FIREBASE_PROJECT_CONFIG && JSON.parse(process.env.FRAMEWORKS_FIREBASE_PROJECT_CONFIG);
+const FRAMEWORKS_APP_OPTIONS = process.env.FRAMEWORKS_APP_OPTIONS && JSON.parse(process.env.FRAMEWORKS_APP_OPTIONS);
 
 const adminApp = initializeAdminApp();
 const adminAuth = getAdminAuth(adminApp);
@@ -40,7 +40,7 @@ const mintCookie = async (req: Request, res: Response) => {
         if (new Date().getTime() / 1_000 - verifiedIdToken.auth_time > ID_TOKEN_MAX_AGE) {
             res.status(301).end();
         } else {
-            const cookie = await adminAuth.createSessionCookie(idToken!, { expiresIn: COOKIE_MAX_AGE }).catch(e => {
+            const cookie = await adminAuth.createSessionCookie(idToken!, { expiresIn: COOKIE_MAX_AGE }).catch((e: any) => {
                 console.error(e.message);
             });
             if (cookie) {
@@ -56,26 +56,26 @@ const mintCookie = async (req: Request, res: Response) => {
 };
 
 const handleAuth = async (req: Request) => {
-    if (!FIREBASE_PROJECT_CONFIG) return;
+    if (!FRAMEWORKS_APP_OPTIONS) return;
     const cookies = cookie.parse(req.headers.cookie || '');
     const { __session } = cookies;
     if (!__session) return;
-    const decodedIdToken = await adminAuth.verifySessionCookie(__session).catch(e => console.error(e.message));
+    const decodedIdToken = await adminAuth.verifySessionCookie(__session).catch((e: any) => console.error(e.message));
     if (!decodedIdToken) return;
     const { uid } = decodedIdToken;
     let app = firebaseAppsLRU.get(uid);
     if (!app) {
-        const isRevoked = !(await adminAuth.verifySessionCookie(__session, true).catch(e => console.error(e.message)));
+        const isRevoked = !(await adminAuth.verifySessionCookie(__session, true).catch((e: any) => console.error(e.message)));
         if (isRevoked) return;
         const random = Math.random().toString(36).split('.')[1];
         const appName = `authenticated-context:${uid}:${random}`;
-        app = initializeApp(FIREBASE_PROJECT_CONFIG, appName);
+        app = initializeApp(FRAMEWORKS_APP_OPTIONS, appName);
         firebaseAppsLRU.set(uid, app);
     }
     const auth = getAuth(app);
     if (auth.currentUser?.uid !== uid) {
         // TODO(jamesdaniels) get custom claims
-        const customToken = await adminAuth.createCustomToken(uid).catch(e => console.error(e.message));
+        const customToken = await adminAuth.createCustomToken(uid).catch((e: any) => console.error(e.message));
         if (!customToken) return;
         await signInWithCustomToken(auth, customToken);
     }
@@ -83,11 +83,11 @@ const handleAuth = async (req: Request) => {
     req.currentUser = auth.currentUser;
 };
 
-export const ssr = onRequest(HTTPS_OPTIONS, async (req: Request, res: Response) => {
+export const handleFactory = (frameworkHandle: (req: Request, res: Response) => void) => async (req: Request, res: Response) => {
     if (req.url === '/__session') {
         await mintCookie(req, res);
     } else {
         await handleAuth(req);
         frameworkHandle(req, res);
     }
-});
+};
