@@ -13,7 +13,6 @@
 // limitations under the License.
 
 import { exec as execCallback, spawn as spawnCallback, ExecOptions, SpawnOptionsWithoutStdio, spawnSync } from 'child_process';
-import { join } from 'path';
 
 export const exec = (command: string, options: ExecOptions={}) => new Promise((resolve, reject) =>
     execCallback(command, options, (error, stdout) => {
@@ -59,10 +58,8 @@ export type DeployConfig = {
 
 export type PathFactory = (...args: string[]) => string;
 
-export const getProjectPathFactory = (config: DeployConfig): PathFactory => (...args) => join(process.cwd(), config.prefix ?? '.', ...args);
-
 export const findDependency = (name: string, cwd=process.cwd()) => {
-    const result = spawnSync('npm', ['list', name, '--json', '--omit', 'dev'], { cwd });
+    const result = spawnSync(Commands.NPM, ['list', name, '--json', '--omit', 'dev'], { cwd });
     if (!result.stdout) return undefined;
     const json = JSON.parse(result.stdout.toString());
     const search = (searchingFor: string, dependencies={}): any => {
@@ -74,4 +71,10 @@ export const findDependency = (name: string, cwd=process.cwd()) => {
         return null;
     }
     return search(name, json.dependencies);
+};
+
+export const Commands = {
+    NPM: process.platform === 'win32' ? 'npm.cmd' : 'npm',
+    next: process.platform === 'win32' ? 'next.cmd' : 'next',
+    ng: process.platform === 'win32' ? 'ng.cmd' : 'ng',
 };
