@@ -1,52 +1,47 @@
-import { spawnSync } from 'child_process';
-import { replaceInFile } from 'replace-in-file';
-import { readJSON } from 'fs-extra';
+import { copyFile } from 'fs/promises';
+
+export {};
+
+const { default: { replaceInFile } } = await import('replace-in-file');
+const { default: { readJSON } } = await import('fs-extra');
 
 const LOCAL_NODE_MODULES = [
-    '@angular-devkit/core/node',
-    '@angular-devkit/core',
-    '@angular-devkit/architect/node',
-    '@angular-devkit/architect',
-    'next/dist/build',
-    'next/dist/export',
-    'next/dist/trace',
-    'nuxt',
+    // Angular
+    '@angular-devkit/core/node/index.js',
+    '@angular-devkit/core/src/index.js',
+    '@angular-devkit/architect/node/index.js',
+    '@angular-devkit/architect/src/index.js',
+    // Next.js
+    'next/dist/build/index.js',
+    'next/dist/export/index.js',
+    'next/dist/trace/index.js',
+    'next/dist/server/config.js',
+    'next/constants.js',
+    // Nuxt v2
+    'nuxt/dist/nuxt.js',
+    '@nuxt/generator/dist/generator.js',
+    '@nuxt/builder/dist/builder.js',
+    // Nuxt3
     '@nuxt/kit/dist/index.mjs',
-    'webpack',
-];
-
-const ES_MODULES = [
-    ['@nuxt/kit', 'dist/index.mjs'],
 ];
 
 const main = async () => {
+
     await replaceInFile({
         files: 'dist/**/*',
-        from: ES_MODULES.map(([path]) =>  new RegExp(`Promise\\.resolve\\(\\)\\.then\\(\\(\\) => __importStar\\(require\\('${path}'\\)\\)\\)`, 'g')),
-        to: ES_MODULES.map(([path, file]) => `import('${path}${file ? `/${file}` : ''}')`),
+        from: LOCAL_NODE_MODULES.map(mod => new RegExp(`require\\(["'\`]${mod}["'\`]\\)`, 'g')),
+        to: LOCAL_NODE_MODULES.map(mod => `require(\`\${pathToFileURL(getProjectPath())}/node_modules/${mod}\`)`),
     });
 
     await replaceInFile({
         files: 'dist/**/*',
-        from: 'Promise.resolve().then(() => __importStar(require(`${process.cwd()}/index.mjs`)))',
-        to: 'import(`${process.cwd()}/index.mjs`)',
-    });
-
-    await replaceInFile({
-        files: 'dist/**/*',
-        from: LOCAL_NODE_MODULES.map(mod => new RegExp(`require\\(["']${mod}["']\\)`, 'g')),
-        to: LOCAL_NODE_MODULES.map(mod => `require(\`\${process.cwd()}/node_modules/${mod}\`)`),
-    });
-
-    await replaceInFile({
-        files: 'dist/**/*',
-        from: LOCAL_NODE_MODULES.map(mod => new RegExp(`import\\(["']${mod}["']\\)`, 'g')),
-        to: LOCAL_NODE_MODULES.map(mod => `import(\`\${process.cwd()}/node_modules/${mod}\`)`),
+        from: LOCAL_NODE_MODULES.map(mod => new RegExp(`import\\(["'\`]${mod}["'\`]\\)`, 'g')),
+        to: LOCAL_NODE_MODULES.map(mod => `import(\`\${pathToFileURL(getProjectPath())}/node_modules/${mod}\`)`),
     });
 
     const { devDependencies } = await readJSON('package.json');
     const from = ['__FIREBASE_FRAMEWORKS_VERSION__'];
-    const to = [`file:${process.cwd()}`];
+    const to = [`file://${process.cwd()}`];
     for (const [dep, version] of Object.entries<Record<string, string>>(devDependencies)) {
         from.push(`__${dep.toUpperCase().replace(/[^A-Z]/g, '_')}_VERSION__`);
         to.push(version as any);
@@ -56,6 +51,8 @@ const main = async () => {
         from,
         to
     });
+
+    copyFile('src/tools.cjs', 'dist/tools.cjs');
 }
 
 main();
